Fixing file names to match tags
