+++ genai-perf/genai_perf/record/types/energy_consumption_base.py
