+++ genai-perf/genai_perf/record/types/energy_consumption_p75.py
@@ -14,20 +14,20 @@
 
 from functools import total_ordering
 
-from genai_perf.record.types.gpu_energy_consumption_base import GPUEnergyConsumptionBase
+from genai_perf.record.types.energy_consumption_base import GPUEnergyConsumptionBase
 
 
 @total_ordering
-class GpuEnergyConsumptionP99(GPUEnergyConsumptionBase):
+class GpuEnergyConsumptionP75(GPUEnergyConsumptionBase):
     """
-    A record for p99 GPU Energy Consumption metric
+    A record for p75 GPU Energy Consumption metric
     """
 
-    tag = GPUEnergyConsumptionBase.base_tag + "_p99"
+    tag = GPUEnergyConsumptionBase.base_tag + "_p75"
 
     def __init__(self, value, device_uuid=None, timestamp=0):
         super().__init__(value, device_uuid, timestamp)
 
     @classmethod
     def header(cls, aggregation_tag=False) -> str:
-        return "p99 GPU Energy Consumption (MJ)"
+        return "p75 GPU Energy Consumption (MJ)"
