+++ genai-perf/genai_perf/record/types/total_gpu_memory_base.py
